requirements.txt
bottle
youtube-transcript-api
requests
waitress
gunicorn
script-reporter
python-dotenv